make tab container tests use driver.WalkObjectTree for tree inspection
This moves the test into the `widget_test` package, because `driver`
depends on `widget`.
This has the advantage that no internal interface is tested.

The position computation for the layout test doesn't modify objects
anymore.

The `ApplyTheme` test is now named correctly (it tests the renderer)
and no longer relies on the internal structure of the renderer.