Always execute certain event callbacks serially
This is to fix crashes and inconsistent behavior when events come in
rapidly. If you went to the multiline form widget, for example, and
mashed some keys and the return key very quickly, you could easily make
it crash because sometimes the key handler and rune handler happened at
the same time, causing a data race.

I'm including mouse events because I imagine people hitting shift and
clicking quickly, or just plain clicking quickly and the mouse down event
happening after (or at the same time as) the mouse up event.