Fix wrong rendering and even crash on appending tabs
The divider and tab bar is appended to tabContainerRenderer.objects, but it
doesn't get updated properly when tabs are added via Append. When appending tabs
the divider and the tab bar are overwritten. When appending 3 or more tabs, it
crashes.

This bugfix removes divider and tab bar from that objects field, and populate it
correctly on appending and removing.