Make ShowConfirm(), ShowInformation() and ShowError return a Dialog
This lets the caller get a pointer to the dialog without having to
increase the API surface.

Although the change is not fully backwards compatible, the amount of
broken client code should be minimal, if any. The only case where
the API does cause a meaningful change is if someone was passing
these changed functions to other functions.