Add support for -appVersion and -buildNumber in fyne package
For macOS, -appVersion is mapped to CFBundleVersionShort while
-buildNumer is used for CFBundleVersion. If no CFBundleVersion
is provided, it's set to the same value as CFBundleVersionShort.

For Windows, -appVersion is expanded to 3 components, then
-buildNumber (which must have 1 component at most) is appended
to it to form the 4 component assembly version that Windows
expects. If no -buildNumber is provided, we append a zero to
get to 4 components.

No other platforms use these 2 new flags for now.