change menu layout test to use rendered images, only
This includes a fix to keep a menu item highlighted
if its submenu is visible.
It also fixes the position handling of a submenu which only
worked if the size and position of the menu item as well as
the size of the canvas were known when the menu item’s
renderer was created.