Add -appVersion and -appBuild parameters to specify versioning
Andorid, darwin and iOS will use version for the display version and build for the internal version id.
Windows will use a combined version which is <version>.<build>, i.e. 1.0.0.1.
These fields are optional for packaging (1.0 and 1 will be used like before otherwise)
but required when building a release (obvs!)