Switch bundled resources to strings
The gopls team has received reports of high memory usage in the Fyne
project. The problem is the very large byte slice literals in the theme
package, which consume about 200MiB for the AST and another 200-400 for
type information.

While this is obviously a deficiency in gopls, it's not easy to fix
while maintaining full functionality. Changing the byte slices to
strings is very easy.

As a further optimization, it might be good to change them to be *used*
as strings, since byte slices have to live on the heap and therefore
consume actual memory, whereas strings can be left on disk until
accessed.