move Child, Exists, and Destroy into URI interface
The underlying work to provide implementations of these methods has not
been done yet pending discussion, so this will break the build.

This also adds Destroy() as described in #1509, though again no
implementation is defined yet.

This also adds detailed comments to the members of the URI interface, to
make it clear what is expected of URI implementations.