fix menu item refreshing
Before the removal of background colors, the refresh
on a menu item for updating the background color
was not really necessary because the menu refresh
already triggered the repaint of the menu (sub)tree.

With the background being normal canvas objects,
two errors in the refresh of the items revealed
which were hard to tackle:
- the previously active item was refreshed _before_
  it was deactivated
  -> the background color did not change
- the newly active item was not refreshed at all
  -> the background color did not change

This led to weird behavior where the items changed
there background to active when becoming inactive
and then never again.