revert the removal of menuBarItem’s hover state
It is useful for inactive menu bars to give
visual feedback. This also pleases the tests :).

However, the behavior is slightly changed:
The hover state is no longer shown if the menu bar
is active and the active item was changed with keyboard
so that the mouse no longer hovers the active item.