Remove custom bundler scripts and rebundle with latest fyne bundle
This removes the `gen.go` custom bundle scripts that were used in `cmd/fyne_demo` and `cmd/fyne_demo` in favour of using `go:generate` in combination with `fyne bundle`.
The only remaining custom bundle script is the one in the theme package, but that requires a lot more work due to the fonts combined with the ammount of icons.