Changes to improve ci
This updates the static analysis to use Go 1.16 for the improvements with using `go install` with a version number (only used for staticcheck for now) and without modifying `go.mod`.

It also switches to a faster go setup action which is about three times faster while also allowing us to see inline staticcheck reports in PR diffs. I also made sure to tell staticcheck what Go version we are targeting.