Some minor code refactor and cleanup
Supercedes #1888

This is a continuation of #1888 with the goal of applying the same fundamental ideas while doing some further manual refactor to clean up the code even more.
It also applies on top of the develop branch so that we edit the correct files while at the same time not including the controversial DeepSource analysis.