Optimizations for AppTabs and DocTabs
This basically does two things. Firstly, we pre-allocate the slice of menu items that we are shoring in the drop-downs.
Secondly, we make sure to use the index of the tab instead of the tab item, for selection.
This avoids ranging through all the items to find the correct index.