Update windows.go to fix miss-clicks issues/2484
From the code above I understand that there is some logic to alter the w.mouseDragStarted state. However in the last if, we use the mouseDragStarted value which we only set once at the beginning of the function. As part of this change, I've updated mouseDragStarted back to false to properly reflect the correct state