Correct mention of Go 1.12 as lowest supported version
The issue #2559 made me aware that our documentation in the README never got updated to reflect Go 1.14 as the official lowest supported version.
I am aiming this twoards develop since 2.1.x (on master) actually had any changes made that would make the compile fail on Go 1.12 (those are only in 2.2.0 afaik).
If people feels like this should be on master as well (which might or might not be necessary), we can simply cherry-pick it over.