Fix shortcuts not working in non-ASCII keyboard layouts
When user pressed the physical keys corresponding to Ctrl+V while using,
for example, Russian keyboard layout, they usually expect a Paste shortcut
to trigger. This is how Windows, Mac, and good graphics toolkits on Linux
work. In fyne this used to be represented as Ctrl+Unknown key, but now the
non-ascii keys are translated when pressed with Ctrl or Alt modifier.