[internal/painter] add test for compositeFace#Glyph(…)
This adjusts the behaviour to comply with the specification of
golang.org/x/image/font.Face#Glyph and return !ok if the face
does not contain the glyph.

This changes the behaviour of faces which return a placeholder
instead. Now the composite face no longer returns placeholder characters.
This will be addressed by a follow-up PR which will add
replacement character rendering to the font drawer.