painter/gl: createProgram helper
This one pretends to use a filesystem, i.e. it gets a filename (without extension) and creates an OpenGL program from the shaders for this filename (<filename>.frag and <filename>.vert). This is an interface like we would have if we already had the go:embed functionality to our hands.