[widget] inline menuItemRenderer#itemPadding()
Except for minSize computation it was never used as it is.
Instead, the layout code “interpreted” it and took fractions of it for
computation.
This indirection hides the layout’s intention and makes the code hard to
read.

The inlining of this method is the first step to a more clean approach
to the menu item layout.