[internal/painter] split PaintImage()
There’s one version which respects the image’s fill mode and might not
render the image at all because it changes the image’s min size which
will trigger a re-draw in the next frame.
The other version is for plain rendering of an image to a certain size
(without a canvas, then).