[internal/painter] PaintImage ignores fill mode when no canvas is given
This reflects the usage better:
For actual rendering (if a canvas is given), the fill mode is respected.
It might happen that the image is not rendered but its min size corrected, then.

For rendering an image for other purposes (no canvas given), the fill mode
is ignored and the image is always rendered to the requested size.