fix scrolling of objects larger than canvas
PR #2893 included an optimization to not paint an object if the
intersection of the visible area and the objects’s current position and
size was empty.
This computation does not consider that the object might have been
scrolled. In effect, the background of large pop-ups (like the icon list
in the demo app) vanished when scrolling down too much.

The optimization is removed for now. If it will ever be re-introduced,
it has to consider scrolling and there have to be tests, of course.