feat: create new folder from file/folder dialog (#3178)
This PR introduces the ability to create new folder directly from the file/folder dialog. This is motivated by the hassle of having to create a folder separately before being able to select it via the folder dialog.

Changes were made to makeUI() to include a create folder button in the options container which would open up a entry form dialog for user to input the new folder name before executing the mkdir syscall. Directory is then refreshed to enable user to select the newly created folder.

Fixes #3174