Allow for doctabs.CreateTabs to take its time before returning a value.
This is a likely use case of showing up a dialog.NewForm and wait for
the result. The problem is that, at least on linux, the input thread is
mixed with the rendering thread. So when a button is tapped and the
interaction is blocked, it does block the main loop.

Another potential fix for this would be to make the Tapped callback
happen in a goroutine too.