Remove unecessary and excessive lock.
Walking the tree already trigger a large number of lock on the widget it
walk. I didn't knew this when I added this unecessary lock. This
excessive locking could lead to a deadlock where a lock inside the walk
was already taken outside the lock and the big lock prevent things from
moving any further.