[internal/driver/glfw] adjust master XML files
The padding seems to have changed (from 4 to 6).
Also, the text size has changed.
For some rectangles the stroke color has changed to default. This does
not affect the test case which is about visibility of a rectangle filled
with the hover color.

With the minor changes to dimensions and positions the mouse movements
of the test are still sufficient.