[internal/driver/glfw] fix several GLFW test issues on Linux
The root cause seems to be some race/concurrency issue with the window
manager. I was not able to fix it properly but waiting for the expected canvas size
with an explicit resize every now and then is a sufficient workaround.

Note: The window size in the `Test_gLDriver_AbsolutePositionForObject`
test was adjusted to a more realistic value. At least the width was not
sufficient. This was not a real issue because the window got resized but
its better to choose a size which is sufficient (and verify it) because
this avoids any surprises.