dialog: Add HideWithSubmit() to form dialog
This allows a form dialog to be hidden and submitted if all validators are passing. The new API cna be very useful when implementing an entry that should close the dialog when submitted. Just calling hide will run the OnClosed() with false in that case so the user will not be able to submit, even if the validation is passing.

Unblocks https://github.com/Jacalz/rymdport/issues/55