Fix potential race condition in preference change listeners
This fixes a sync.WaitGroup instance that was shared in the struct and thus potentially across multiple threads.
A user could get weird behaviour by having another call of the function modify the same variable.
As a second benefit, we also get rid of an unnecessary struct parameter (it ddid not need to be there even without the race condition).