Move away from struct creation using new() when we have access to fields
It seems like we (and many other projects but that's just speculation) are moving away from using new() when creating structs (using new(int) still is useful). My opinion is that t is a lot clearer to create the struct directly, without using new(), and we also get direct field access which can decrease inlining costs and sometimes allow the compiler to optimize things a bit more. This changes a few places where it seemed to make sense to not use new().