FileDialog view layout preference (#4278)
Prior to this change, the FileDialog always started with grid view regardless of the previously selected view. This pull request adds a `fileDialogViewLayout` key to app preferences to save the last view layout selected. The next time a FileDialog is show it will start with the last view layout selected by the user.