systray: ensure title is not empty
Some desktop environments like the one provided by Ubuntu 22.04 requires the systray title to be not empty to be displayed.
This commit ensure the title is always set trying to use the app name from metadata. If the app name is not set will fallback to the app ID

Fixes #3678