Reduce string allocations in mimetype code
The code handling mimetypes was using strings.Split() to split the string and that allocates. We can use strings.IndexByte() to get an index and subslice instead. I also refactored some code to avoid duplicating mimetype split logic.

The new internal/repository/mime package exists to avoid cycling imports.

NOTE: The split logic is slightly different because a second / would be includes as part of the subtype now but as far as I can see from reading the spec (https://specifications.freedesktop.org/shared-mime-info-spec/shared-mime-info-spec-0.21.html) it should not be a problem as the format is "mainType/subType;weight" so there should not be any more separators.