Update go-gl/glfw to v3.3.10 and remove workaround
This is a great GLFW update (see https://github.com/go-gl/glfw/pull/396) because it not only fixes various bugs but one of the bugs that it fixes addresses a bug that we have been working around for a long time.

The workaround for this issue involved defering a recover() to catch the panic that could occur for incorrectly invalid keycodes. This did not only degrade performance due to having to run recover but the defer was also making it impossible to inline the method. Given that we are running this in the (rather hot) runloop, I expect the perofrmance improvement to be quite substancial.