Update portal to v0.2.0 and implement varipus code improvements
This removes some dead code, fixes a possible race condition when adding notifications and contains and improved watcher for settings changes. The latter is likely a performance improvement when changing theme because the old one would match both the internal Gnome key and the standard value and thus refreshing the window twice.