Don't create the renderer in the constructor for infinite progressbar
This fixes the infinite progress bar code to not create the renderer inside the constrcutor. The old behaviour was not only a potential performance and memory leak and just seems generally incorrect from a best practices standpoint.

The tests had to be modified to actually render the object so this is a change in behaviour but I would not call it breaking given that I consider the old behaviour a incorrect and a bug.