[internal/driver/mobile] separate chrome box offset from its size in mobileCanvas#sizeContent
This simplifies the next diff where the offset computation is extracted.
This changes the local decision for the full-width window-head
from “chrome box contains more than one element”
to “chrome box displaces content”.
This is still not a total straight forward decision but not worse then
the old one which will survive in the extracted offset computation.