[internal/driver/mobile] flat out  canvas’ properties
The grouping “rule” was hard to comprehend and does not really help
when trying to understand which properties are there and for which purpose.

If a grouping is really wanted, it should be achieved by prefixing property
names like the already existing `onTyped`, `lastTap` and `drag`.
This also makes clear which context a property belongs to when stumbling
over it somewhere in the code.