[theme] introduce (yet unused) “on” colors for error, success and warning
Besides “primary” these are the colors used to express different
importance levels. To achieve a reasonable contrast, the color to paint
on these (or to paint these on) might be different to the default
foreground (or background) color.
The new “on” colors are introduced with the foreground value but
probably change, soon.