[app][internal/app] fix build flags
XDG files should not be included in mobile. For non-Android non-IOS
mobile builds (e.g., running tests with `-tags mobile`) the `app_other.go`
has to be included, to provide `OpenURL` and `SendNotification`.

The whole scheme has clearly evolved over time.
Might be useful to separate the different functionalities and get a more
clear scheme in the end.