[theme] move implementation of .PrimaryForegroundColorNamed() to internal/theme
Thus, it can be shared with the settings app even if the theme package
method gets deprecated.

This also introduces separated color values for the various variants
which are copies of the background colors used before.

And last but not least, the internal method is named correctly. The public one will vanish soon.