Make the large scrollbar slightly narrower
It was looking a bit too big previously and did not fit in well with the
rest of the UI because of it. This makes it smaller, now four times as
big as the small bar. We might want to make it smaller in the future but
this seems like a step in the right direction.