Fix color lookup not working as expected on Linux
The theme value was looked up before being set for the first time. I also updated to a new version of the portal package that lets us not only use the value when the theme changes (instead of manually loading it again) but also falls back to use the old Read function to look up settings if ReadOne fails.

Should hopefully fix #5029