added a search bar to the file dialog
the search bar takes advantage of the existing filtering system in order
to only show files or filders that match a specific prompt. for now,
there is no way to recursively search through files, or use something
like regex, though that could be added as options later.

the main internal change needed is, in order to be backwards compatible,
having a list of filters instead of just a single filter at a time. that
way, library users can still have many filters while including the
search bar.