driver/glfw: Merge waitForStart and done channels
With the threading changes, we no longer need to have separate
channels for done and waitForStart given that showing a window no
longer is expected to happen on the wrong thread. This allows
allows us to first wait for starting and then reopen the channel
to wait for shutting down.