Remove atomic.Pointer from fyne.CurrentApp()
The main benefit here is that someone using the race detector will know
that they are using fyne.CurrentApp() from the wrong goroutine in a
migrated app and it should be safe for non migrations given the fact
that the app is set once only. Another benefit is of course that
frequest app lookups should be much faster (not only are we avoiding the
atomic lookup but fyne.CurrentApp() is now also inlineable).