Better solution for faster internal maps when migrated
This uses a lock that turns off when migrated instead of migrating the whole async map. Firstly, this avoids potentialy confusion with async.Map potentially not being safe when it should be, Secondly, it avoids the issue with generics not supporting comparable map keys on Go 1.19 so should result in a performance improvement for all maps.