Fix issues from new gofumpt release
This fixes code formatting now that the latest gofumpt doesn't want bare return statements. I think this is a good thing as it makes code more readable while still allowing named return values. From the code lines changed in this PR, it also points to it being more consistent with the rest of our code.