Disregard a fake "Alt key up" event sent by glfw after unfocusing
This used to trigger the main menu of the application after
user pressed Alt+Tab: the application receives an "alt key
pressed" event (as it should), then it receives a "window
focused" event, and then glfw immediately manufactures a
fake "alt key released" event (even if the user continues
to hold the Alt key). Previously the unfocusing callback did
not update the internal variables used to keep track of key
presses, so this sequence of events got interpreted as
"user pressed and then released the Alt key", thus opening
the main menu.